requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
astroid==2.4.2
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-filter==2.4.0
django-periodically==0.3.0
djangorestframework==3.12.4
docopt==0.6.2
Faker==8.0.0
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
Js2Py==0.71
lazy-object-proxy==1.4.3
legacy==0.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
packaging==20.9
Pillow==8.2.0
pipwin==0.5.1
pycodestyle==2.7.0
pyjsparser==2.7.1
pylint==2.6.0
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
serializers==0.2.4
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
